# Fix correctness and performance regression in `require`