# Description:

A recent fix in rubygems, #1868, caused both a perfomance and a functionality regression in our custom `require`.

## What was the end-user or developer problem that led to this PR?

On one hand, `require` was made slower because it traverse all activated gems by default.

On the other hand, the unintended manual traversing of those paths also caused divergences in functionality from the standard `require` since the `$LOADED_FEATURES` cache would no longer be hit.

## What is your fix for the problem, implemented in this PR?

This PR keeps track of the activated gem paths in order to traverse only directories prepended to the `$LOAD_PATH` by ruby `CLI` flags or environment variables.

Fixes #3491.
Fixes #3631.
Fixes #3647.

# Tasks:

- [x] Describe the problem / feature
- [x] Write tests
- [x] Write code to solve the problem
- [ ] Get code review from coworkers / friends

I will abide by the [code of conduct](https://github.com/rubygems/rubygems/blob/master/CODE_OF_CONDUCT.md).