@lamont-granquist 2020-05-20
Timing tests:

current 3.1 behavior:  /opt/chef-workstation/bin/kitchen list  11.60s user 4.53s system 54% cpu 29.468 total
completely removed behavior:  /opt/chef-workstation/bin/kitchen list  2.58s user 0.82s system 48% cpu 7.046 total
this patch's behavior:  /opt/chef-workstation/bin/kitchen list  2.57s user 0.82s system 48% cpu 7.053 total

So that looks good.

I'm also seeing that it accurate:

Throwing an `-I/foo/bar` in the binstub, I'm only seeing these two being looped over first:

```
/opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/test-kitchen-2.5.1/bin/../lib
/foo/bar
```

The code reads correct to me as well.