@lamont-granquist 2020-05-21
Looks even faster:

/opt/chef-workstation/bin/kitchen list  2.41s user 0.62s system 48% cpu 6.201 total