@deivid-rodriguez 2020-05-21
Yeah! Now we no longer do the `-I` paths traversal for every require, but only when a name present on a default gem is required :)